// client/src/app/enums.ts
export
enum SIDEBAR_ICON_INDEX {
  SELECTION = 0,
  DRAWING = 1,
  SHAPE = 2,
  COLOR = 3,
  GRID = 4
}

export
enum EXTERNAL_SERVICES_INDEX {
  DRAWBOARDSERVICE = 0,
  DATASERVICE = 1,
  DRAWINVOKERSERVICE = 2,
  SELECTIONSERVICE_INDEX = 3,
  GRIDSERVICE = 4,
  EXPORTDRAWSERVICE = 5
}

export
enum MODALES_INSTANCES_INDEX {
  ATTRIBUTEFOCUS_INDEX = 0,
  SHOWCREATEDRAW_INDEX = 1,
  SHOWSAVEDRAW_INDEX = 2,
  SHOWEXPORTDRAW_INDEX = 3,
  SHOWGALLERY_INDEX = 4
}

export
enum LAST_SELECTED_TOOL_INDEX {
  DRAWINGTOOL_INDEX = 0,
  SHAPETOOL_INDEX = 1,
  COLORTOOL_INDEX = 2
}

export
enum ICON_STATE_INDEX {
  SIDEBAR_STATE = 0,
  DRAWTOOLSTATE_INDEX = 1,
  SHAPETOOLSTATE_INDEX = 2,
  COLORTOOLSTATE_INDEX = 3
}

export
enum SELECTION_MENU_INDEX {
  SIDEBAR_INDEX = 0,
  ACTIVETOOL_INDEX = 1
}

export
enum DRAWING_ICON_INDEX {
  PENCIL = 0,
  BRUSH = 1,
  SPRAY = 2,
  ERASER = 3,
  APPLICATOR = 4
}

export
enum SHAPE_ICON_INDEX {
  LINE = 0,
  RECTANGLE = 1,
  ELLIPSE = 2,
  POLYGON = 3,
}

export
enum COLOR_ICON_INDEX {
  BUCKET = 0,
  PICKER = 1,
  APPLICATOR = 2
}

export
enum TOOLS_INDEX {
  SELECTION = 0,
  PENCIL = 1,
  BRUSH = 2,
  ERASER = 3,
  LINE= 4,
  RECTANGLE = 5,
  ELLIPSE = 6,
  POLYGON = 7,
  SPRAY = 8,
  PICKER = 9,
  GRID = 10,
  APPLICATOR = 11,
  BUCKET = 12
}
export
enum SHAPE_TYPES {
  FILL = 'fill',
  CONTOUR = 'contour',
  CONTOUR_AND_FILL = 'none'
}